fix: Update the constraints of the MapView on LocationPhotos controller
Attach the top constraint of the mapView to the superview top,
while updating the collection view top constraint to be attached to
the bottom of the map view (0).
Prevent the app to be used on landscape mode as this causes
blank space between the map and the collection despite the constraint.